fix: 파라매터 타입 Integer -> Long 변경

+++ src/main/java/christmas/domain/event/GiftEvent.java
@@ -2,9 +2,9 @@ package christmas.domain.event;
 
 import christmas.domain.menu.MenuCategory;
 
-public class GiftEvent implements Event<Integer, MenuCategory, Integer>{
+public class GiftEvent implements Event<Long, MenuCategory, Long>{
     @Override
-    public boolean support(Integer OrderAmountBeforeDiscount) {
+    public boolean support(Long OrderAmountBeforeDiscount) {
         if (OrderAmountBeforeDiscount > EventConfig.MINIMUM_ORDER_AMOUNT_FOR_GIFT_EVENT ) {
             return true;
         }
@@ -12,7 +12,7 @@ public class GiftEvent implements Event<Integer, MenuCategory, Integer>{
     }
 
     @Override
-    public MenuCategory discount(Integer amount) {
+    public MenuCategory discount(Long amount) {
         return MenuCategory.CHAMPAGNE;
     }
 }
